v1.7.6_161020   为了更好地兼容系统，我们建议使用PHP5.3以上环境运行WSTMall。后期我们新增功能主要依开发计划以及参考用户在本网站提出的建议，希望更多用户朋友能通过本站向我们提出更多的宝贵意见。
本次版本要新增以下功能：

1.修改全国配店铺的商品可在列表显示
2.修复商品详情页右侧购物车商品较多时不能下拉
3.修复货到付款，也会超库存下单bug
4.修复含规格商品 加减购买数量会以推荐的那个规格来限制bug
5.兼容360浏览器图片上传问题
6.系统代码优化

*若已安装的用户请参照以下步骤：
1、wstmall/Upgrade/Upgrade覆盖升级。

WSTMall更新计划可参考：http://bbs.wstmall.net/forum.php?mod=viewthread&tid=141

WSTMall在成长的脚步上离不开大家的热心支持，大家有什么意见或者建议可通过以下渠道反馈给我们：
WSTMall用户交流群：50056937
WSTMall客服QQ：707563272
WSTMall客服电话：020-85289921
WSTMall邮箱： wstmall@qq.com
WSTMall官网：http://www.wstmall.net
百度贴吧：http://tieba.baidu.com/wstmall
Github：https://github.com/wstmall/wstmall.git
Oscgit：https://git.oschina.net/wstmall/wstmall.git

+++ Apps/Admin/Model/AdsModel.class.php
@@ -3,7 +3,7 @@
 /**
  * ============================================================================
  * WSTMall开源商城
- * 官网地址:http://www.wstmall.com 
+ * 官网地址:http://www.wstmall.net
  * 联系QQ:707563272
  * ============================================================================
  * 广告服务类
@@ -71,7 +71,7 @@ class AdsModel extends BaseModel {
 	  * 分页列表
 	  */
      public function queryByPage(){
-     	$adPositionId = (int)I('adPositionId');
+     	$adPositionId = I('adPositionId');
      	$adDateRange = I('adDateRange');
      	$adName = WSTAddslashes(I('adName'));
 	 	$sql = "select a.*,a1.areaName areaName1,a2.areaName areaName2,a3.areaName areaName3,c.communityName 
@@ -79,7 +79,7 @@ class AdsModel extends BaseModel {
 	 	        left join __PREFIX__areas a2 on a.areaId2 = a2.areaId 
 	 	        left join __PREFIX__areas a3 on a.areaId3 = a3.areaId 
 	 	        left join __PREFIX__communitys c on a.communityId = c.communityId where 1=1 ";
-	 	if($adPositionId!="")$sql.="  and adPositionId=".$adPositionId;
+	 	if($adPositionId!="")$sql.="  and adPositionId=".(int)$adPositionId;
 	 	$sql.=' order by adId desc';
 		$rs = $this->pageQuery($sql);
 		if(count($rs[root])>0){
